requirements.txt
streamlit>=1.44.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.0.0
gdown>=5.2.0
clip @ git+https://github.com/openai/CLIP.git
